Removed api from file path;edited obj values to line up with client
