+++ config/keys.js
@@ -1,6 +1,8 @@
-//module.exports = {
- //   mongoURI: 'mongodb+srv://dbuser:<password>@clusterocrush-ljprw.mongodb.net/test?retryWrites=true&w=majority'}
-
- //Tami's mongodb
 module.exports = {
-    mongoURI: 'mongodb+srv://thoff:mOOkie153!@cluster0-dxzz4.mongodb.net/test?retryWrites=true&w=majority' }
+   mongoURI: 'mongodb+srv://dbuser:Password1!@clusterocrush-ljprw.mongodb.net/test?retryWrites=true&w=majority'}
+
+//  //Tami's mongodb
+// module.exports = {
+//     mongoURI: 'mongodb+srv://thoff:mOOkie153!@cluster0-dxzz4.mongodb.net/test?retryWrites=true&w=majority' }
+
+// mongodb+srv://thoff:<password>@cluster0-dxzz4.mongodb.net/test?retryWrites=true&w=majority
